requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
apify_client==1.7.0
apify_shared==1.1.1
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
dataclasses-json==0.6.6
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inexactsearch==1.0.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.0
langchain-core==0.2.0
langchain-text-splitters==0.2.0
langcodes==3.4.0
langsmith==0.1.60
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
marshmallow==3.21.2
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==23.2
preshed==3.0.9
pydantic==2.7.1
pydantic_core==2.18.2
pyspellchecker==0.8.1
PyYAML==6.0.1
requests==2.32.1
setuptools==69.5.1
silpa_common==0.3
smart-open==6.4.0
sniffio==1.3.1
soundex==1.1.3
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
tenacity==8.3.0
thinc==8.2.3
tqdm==4.66.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.4
